tidy(ui): organise postprocessing files
