+++ invokeai/frontend/web/src/features/parameters/components/PostProcessing/ParamPostProcessingModel.tsx
